+++ primitives/state-machine/src/testing.rs
@@ -392,49 +392,4 @@ mod tests {
 		ext.commit_all().unwrap();
 		assert!(ext.backend.eq(&backend), "Both backend should be equal.");
 	}
-
-	#[test]
-	fn execute_and_generate_proof_works() {
-		use codec::Encode;
-		let mut ext = TestExternalities::<BlakeTwo256>::default();
-
-		ext.insert(b"a".to_vec(), vec![1u8; 33]);
-		ext.insert(b"b".to_vec(), vec![2u8; 33]);
-		ext.insert(b"c".to_vec(), vec![3u8; 33]);
-		ext.insert(b"d".to_vec(), vec![4u8; 33]);
-
-		let pre_root = ext.backend.root().clone();
-		let (_, proof) = ext.execute_and_prove(|| {
-			sp_io::storage::get(b"a");
-			sp_io::storage::get(b"b");
-			sp_io::storage::get(b"v");
-			sp_io::storage::get(b"d");
-		});
-
-		let compact_proof = proof
-			.clone()
-			.into_compact_proof::<sp_runtime::traits::BlakeTwo256>(pre_root)
-			.unwrap();
-		let compressed_proof = zstd::stream::encode_all(&compact_proof.encode()[..], 0).unwrap();
-
-		// just an example of how you'd inspect the size of the proof.
-		println!("proof size: {:?}", proof.encoded_size());
-		println!("compact proof size: {:?}", compact_proof.encoded_size());
-		println!("zstd-compressed compact proof size: {:?}", &compressed_proof.len());
-
-		// create a new trie-backed from the proof and make sure it contains everything
-		let proof_check = create_proof_check_backend::<BlakeTwo256>(pre_root, proof).unwrap();
-		assert_eq!(proof_check.storage(b"a",).unwrap().unwrap(), vec![1u8; 33]);
-
-		let _ = ext.execute_and_prove(|| {
-			sp_io::storage::set(b"a", &vec![1u8; 44]);
-		});
-
-		// ensure that these changes are propagated to the backend.
-
-		ext.execute_with(|| {
-			assert_eq!(sp_io::storage::get(b"a").unwrap(), vec![1u8; 44]);
-			assert_eq!(sp_io::storage::get(b"b").unwrap(), vec![2u8; 33]);
-		});
-	}
 }
